Restructured client directories and renamed them
