+++ client/src/App.jsx
@@ -1,17 +1,15 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
 import './App.css';
-import Nav from "./components/Nav/Nav.jsx";
-import Detail from './views/Detail/Detail.jsx';
-import Home from "./views/Home/Home.jsx";
-import Basket from "./views/Basket/Basket.jsx";
-import Login from "./views/Login/Login.jsx";
-import NotFound from "./views/NotFound/NotFound.jsx";
-import MyProfile from "./views/MyProfile/MyProfile.jsx";
-import Footer from "./components/Footer/Footer.jsx";
-import ReduxExperiment from './views/ReduxExperiment';
-import DashBoard from './views/DashBoard/DashBoard.jsx';
-import HomeDB from './views/DashBoard/HomeDB.jsx';
-import EditForm from './views/DashBoard/EditForm';
+import Nav from "./clientComponents/Nav/Nav.jsx";
+import Detail from './clientViews/Detail/Detail.jsx';
+import Home from "./clientViews/Home/Home.jsx";
+import Login from "./clientViews/Login/Login.jsx";
+import NotFound from "./clientViews/NotFound/NotFound.jsx";
+import MyProfile from "./clientViews/MyProfile/MyProfile.jsx";
+import Footer from "./clientComponents/Footer/Footer.jsx";
+import DashBoard from './adminViews/DashBoard.jsx';
+import HomeDB from './adminViews/HomeDB.jsx';
+import EditForm from './adminViews/EditForm';
 
 
 
@@ -26,14 +24,10 @@ function App() {
         <Route path="/" element={<Navigate to="/home" />} />
         <Route path="/login" element={<Login />} />
         <Route path="/myprofile" element={<MyProfile />} />
-        <Route path="/basket" element={<Basket />} />
         <Route path='/dashboard' element={<HomeDB />} />
         <Route path='/dashboard/create' element={<DashBoard />} />
         <Route path='/dashboard/edit/:id' element={<EditForm />} />
         <Route path='/detail/:id' element={<Detail />} />
-
-        <Route path="/reduxexperiment" element={<ReduxExperiment />} />
-
         <Route path="*" element={<NotFound />} />
       </Routes>
       <Footer />
