+++ client/src/adminViews/HomeDB.jsx
@@ -3,10 +3,10 @@ import React from "react";
 import styles from "./Home.module.css";
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import SearchBar from "../../components/SearchBar/SearchBar";
-import { getFoods,setCurrentPageAction } from "../../redux/foodActions.js";
-import CardsContainer from "../../components/DashBoard/CardsContainer";
-import Paginado from "../../components/DashBoard/Paginado";
+import SearchBar from "../clientComponents/SearchBar/SearchBar";
+import { getFoods,setCurrentPageAction } from "../redux/foodActions.js";
+import CardsContainer from "../adminComponents/CardsContainer";
+import Paginado from "../adminComponents/Paginado";
 // import EditForm from "./EditForm";
 import axios from "axios";
 import { Link } from "react-router-dom";
