+++ client/src/clientViews/Detail/Detail.jsx
