+++ client/src/clientViews/Detail/Detail.module.css
