--- client/src/components/Basket/Basket.jsx
@@ -1,9 +1,0 @@
-function Order(){
-    return(
-        <div>
-            <h1>Carrito de compras</h1>
-        </div>
-    )
-}
-
-export default Order;
