--- client/src/views/Basket/Basket.jsx
@@ -1,10 +1,0 @@
-function Basket() {
-    return (
-        <div>
-            NIY: This is the Basket.
-        </div>
-    )
-};
-  
-export default Basket;
-
